move runs_in_parent to exec_utils.c
